usb: dwc3: gadget: disable lpm for rockchip platform
The rockchip DWC3 controllers are LPM capable, but we find
that UVC function compatibility issue on some Windows PCs
if enable LPM capable. And refer to the Logitech USB Cameras,
they're usually disable LPM capable. Considering that the
benefit form LPM capable is insignificant on Rochip platforms,
so we disable the LPM capable by default.

Change-Id: Ibdfb1131b0ecf5f6381f9eb1fc6bea7e7ecb8022
Signed-off-by: William Wu <william.wu@rock-chips.com>
Signed-off-by: Simão Gomes Viana <devel@superboring.dev>