UPSTREAM: net: stmmac: Fixed mtu channged by cache aligned
Since the original mtu is not used when the mtu is updated,
the mtu is aligned with cache, this will get an incorrect.
For example, if you want to configure the mtu to be 1500,
but mtu 1536 is configured in fact.

Fixed: eaf4fac478077 ("net: stmmac: Do not accept invalid MTU values")
Change-Id: I0dc9fee8540fe76d1e583112bbe5326091b8a901
Signed-off-by: David Wu <david.wu@rock-chips.com>
Link: https://lore.kernel.org/r/20210113034109.27865-1-david.wu@rock-chips.com
Signed-off-by: Jakub Kicinski <kuba@kernel.org>
(cherry picked from commit 5b55299eed78538cc4746e50ee97103a1643249c)
Signed-off-by: Simão Gomes Viana <devel@superboring.dev>