PCI: rockchip: dw: Fix unblance pm call for fake probe
We move all the probe stuff to kthread so that it won't block
the system to go on probing other drivers. But that introduced
it bug that PM calls would not be removed by driver core. We
trace back the platform driver to add PM callbacks if we really
detect active devices.

Fixes: 79ac46bdea6b ("PCI: rockchip: dw: Add kthread to probe PCIe devices")
Change-Id: I93a331fe6c17d97b71010780fba4afbc2dbf6216
Signed-off-by: Shawn Lin <shawn.lin@rock-chips.com>
Signed-off-by: Simão Gomes Viana <devel@superboring.dev>